The uploaded pdf has comments
The comments definitely have a theme - "consider your audience."   While I can tell you have done so to some degree, please remember that freshman in Statics won't have studied Kirchoff's equations, unless they had it in high school.  Also, this is the first time that they've used a Jupyter notebook, so more introduction to the notebook itself is needed.  Finally, there are plenty of comments to draw out the ideas you've started to put in the notebook.